MLECO-5340: Initial VSI support
Adding basic Virtual Streaming Interface (VSI) support for HAL camera
interface alongside the default `hal_camera_static_images` library.
Platform drivers for MPS3 and MPS4 targets can enable use of this
library based on a new CMake configuration parameter `FVP_VSI_ENABLED`
By default this is turned OFF.

This will allow all vision-based examples to stream images from camera
connected to a host machine. Although VSI could also be used to read
frames from an input video file on disk (on host), provision for this
has not been added.

+++ source/hal/source/components/npu_ta/ethosu_ta_init.c
@@ -33,7 +33,7 @@ static uint32_t init_ta(uintptr_t base_addr,
     }
 
     ta_set_all(&ta, ta_s);
-    info("Configured TA_%s\t@0x%" PRIx32 "\n", name, base_addr);
+    info("Configured TA_%s\t@0x%" PRIx32 "\n", name, (uint32_t)base_addr);
     return 0;
 }
